Move api middleware into its own file
